+++ API/session.d.ts
@@ -1,0 +1,8 @@
+declare module 'express-session' {
+    interface SessionData {
+      username: string;
+      authorized: boolean;
+    }
+   }
+   
+export {};
