+++ src/components/header/Header.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Header = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default Header;
