--- ts-group-project/API/session.d.ts
@@ -1,8 +1,0 @@
-declare module 'express-session' {
-    interface SessionData {
-      username: string;
-      authorized: boolean;
-    }
-   }
-   
-export {};
