--- ts-group-project/src/style/style.css
@@ -1,104 +1,0 @@
-/* All */
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-}
-header {
-    top: 0;
-    position: absolute;
-    background-color: #0CA14B;
-    padding: 10px;
-    text-align: center;
-    color: white;
-    width: 100%;
-}
-main {
-    background-color: #f5f5f5; /* Light Gray */
-    padding: 2rem; /* Adjust padding for better responsiveness */
-    position: absolute;
-    top: 70%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-align: center; /* Center the text inside the main box */
-    width: 60%;
-}
-
-
-/* Home */
-.article-list {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 20px;
-}
-.article-box {
-    background-color: white;
-    padding: 15px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.homepage-container {
-    align-items: center;
-    width: 100%;
-}
-.logout-container {
-    margin-top: -11vh;
-    display: flex;
-    justify-content: flex-end;
-    align-items: flex-end;
-    flex-grow: 1;
-}
-.logout-button {
-    background-color: #B30000;
-    color: white;
-}
-.logout-button:hover {
-    background-color: #9A0000;
-}
-
-
-/* Login and Signup */
-form {
-    display: flex;
-    flex-direction: column;
-    max-width: 300px;
-    margin: 0 auto;
-}
-div {
-    margin-bottom: 15px;
-}
-a {
-    color: #007BFF;
-    text-decoration: none;
-}
-
-/* FormButton */
-button {
-    background-color: #0CA14B;
-    color: white;
-    padding: 10px 15px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-button:hover {
-    background-color: #00714B;
-}
-div > button {
-    width: 100%;
-}
-
-/* FormInput */
-label {
-    display: block;
-    margin-bottom: 5px;
-}
-input {
-    width: 100%;
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    box-sizing: border-box;
-    margin-bottom: 10px;
-}
